refactor(credentials): fold account and trackers into credentials
`@sidecar/account` and `@sidecar/trackers` are both "who you are / what
authorizes you", both already depended on `@sidecar/credentials`, and shared
one consumer. Both become directories of `credentials`; two packages
disappear, and the two files that are hosted-service clients rather than
credentials move to the packages that own their vocabulary.
