+++ apps/desktop/src/renderer/AGENTS.md
@@ -29,7 +29,7 @@ The same trap arrives through a package barrel, where nothing greps for it.
 Importing `@sidecar/calendar` for one string constant resolves that package's
 whole export graph, `node:http` included. Packages that hold both a vocabulary
 and a Node flow open a door for the vocabulary alone. Import
-`@sidecar/calendar/vocabulary`, `@sidecar/account/snapshot`,
+`@sidecar/calendar/vocabulary`, `@sidecar/credentials/snapshot`,
 `@sidecar/providers/superset/sign-in-stage`, not the barrel.
 
 ## Panel motion
