--- packages/account/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "@sidecar/account",
-  "version": "0.5.0",
-  "private": true,
-  "type": "module",
-  "exports": {
-    ".": "./src/index.js",
-    "./snapshot": "./src/snapshot.js"
-  },
-  "types": "./src/index.ts",
-  "scripts": {
-    "test": "tsx --test 'src/**/*.test.ts'",
-    "typecheck": "tsc --project tsconfig.json"
-  },
-  "devDependencies": {
-    "@types/node": "26.2.0",
-    "tsx": "4.23.12",
-    "typescript": "7.0.2"
-  },
-  "dependencies": {
-    "@sidecar/credentials": "workspace:*",
-    "@sidecar/hosted": "workspace:*",
-    "@sidecar/session": "workspace:*",
-    "@sidecar/settings": "workspace:*",
-    "@sidecar/wire": "workspace:*"
-  }
-}
