--- packages/account/src/index.ts
@@ -1,23 +1,0 @@
-export {
-  AccountClient,
-  AccountClientError,
-  type AccountIdentity,
-  type AccountTokens,
-  type FetchLike,
-  type StoredAccount,
-} from "./client.js";
-export {} from "./deletion.js";
-export { accountGateOpen } from "./gate.js";
-export {} from "./loopback.js";
-export {
-  type AccountPreferencesAnswer,
-  AccountPreferencesClient,
-} from "./preferences.js";
-export { AccountSessionManager } from "./session-manager.js";
-export {
-  ACCOUNT_PROVIDER,
-  ACCOUNT_STATUS,
-  type AccountProvider,
-  type AccountSnapshot,
-} from "./snapshot.js";
-export { HostedVaultClient } from "./vault.js";
