+++ packages/credentials/package.json
@@ -6,6 +6,7 @@
   "exports": {
     ".": "./src/index.js",
     "./loopback-page": "./src/loopback-page.js",
+    "./snapshot": "./src/account/snapshot.js",
     "./vocabulary": "./src/credential-providers.js"
   },
   "types": "./src/index.ts",
@@ -14,6 +15,7 @@
     "typecheck": "tsc --project tsconfig.json"
   },
   "dependencies": {
+    "@sidecar/hosted": "workspace:*",
     "@sidecar/issues": "workspace:*",
     "@sidecar/session": "workspace:*",
     "@sidecar/surface": "workspace:*",
