+++ packages/credentials/src/account/snapshot.ts
