+++ packages/credentials/src/index.ts
@@ -1,4 +1,20 @@
 export {
+  AccountClient,
+  AccountClientError,
+  type AccountIdentity,
+  type AccountTokens,
+  accountGateOpen,
+  type FetchLike,
+  type StoredAccount,
+} from "./account/client.js";
+export { AccountSessionManager } from "./account/session-manager.js";
+export {
+  ACCOUNT_PROVIDER,
+  ACCOUNT_STATUS,
+  type AccountProvider,
+  type AccountSnapshot,
+} from "./account/snapshot.js";
+export {
   CLOUD_AGENT_PROVIDER_LIST,
   CREDENTIAL_CONNECTION,
   CREDENTIAL_PROVIDER_ID,
@@ -13,6 +29,13 @@ export {
   VOICE_CREDENTIAL_PROVIDER_ID,
 } from "./credential-providers.js";
 export { ACCESS_TOKEN_EXPIRY_SLACK_MS } from "./expiry.js";
+export { LinearCredentials } from "./linear/credentials.js";
+export {
+  type LinearGrant,
+  LinearSignIn,
+  linearSignInConfig,
+} from "./linear/oauth.js";
+export { LinearIssueTracker } from "./linear/tracker.js";
 export {
   accountLoopbackPage,
   LOOPBACK_CONNECTION_SOURCE,
