--- packages/trackers/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "@sidecar/trackers",
-  "version": "0.5.0",
-  "private": true,
-  "type": "module",
-  "exports": {
-    ".": "./src/index.js"
-  },
-  "types": "./src/index.ts",
-  "scripts": {
-    "test": "tsx --test 'src/**/*.test.ts'",
-    "typecheck": "tsc --project tsconfig.json"
-  },
-  "devDependencies": {
-    "@types/node": "26.2.0",
-    "tsx": "4.23.12",
-    "typescript": "7.0.2"
-  },
-  "dependencies": {
-    "@sidecar/credentials": "workspace:*",
-    "@sidecar/issues": "workspace:*",
-    "@sidecar/wire": "workspace:*"
-  }
-}
