--- packages/trackers/src/index.ts
@@ -1,7 +1,0 @@
-export { LinearCredentials } from "./linear/credentials.js";
-export {
-  type LinearGrant,
-  LinearSignIn,
-  linearSignInConfig,
-} from "./linear/oauth.js";
-export { LinearIssueTracker } from "./linear/tracker.js";
